pyproject.toml
[project]
name = "superbench-assignment"
version = "0.1.0"
description = ""
authors = [
    {name = "Sean",email = "91415054+sean85120@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.88.0,<2.0.0)",
    "langchain-openai (>=0.3.24,<0.4.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "chromadb (>=0.5.0,<0.6.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "greenlet (>=3.2.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
